Solution to Logical OR Operation

+++ Logical OR Operation@@67882a731b4392d8fc1577dd/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main()
 {
-    int x,y;
-    scanf("%d%d",&x,&y);
+    float x,y;
+    scanf("%f%f",&x,&y);
     if(x||y>0)
     printf("True");
     else
